+++ src/main/java/com/openggf/game/sonic3k/objects/Sonic3kObjectRegistry.java
@@ -18,6 +18,7 @@ import com.openggf.game.sonic3k.objects.badniks.BubblesBadnikInstance;
 import com.openggf.game.sonic3k.objects.badniks.OrbinautBadnikInstance;
 import com.openggf.game.sonic3k.objects.badniks.PenguinatorBadnikInstance;
 import com.openggf.game.sonic3k.objects.badniks.PoindexterBadnikInstance;
+import com.openggf.game.sonic3k.objects.badniks.RibotBadnikInstance;
 import com.openggf.game.sonic3k.objects.badniks.TunnelbotBadnikInstance;
 import com.openggf.game.sonic3k.objects.badniks.TurboSpikerBadnikInstance;
 import com.openggf.game.sonic3k.constants.S3kZoneSet;
@@ -131,13 +132,16 @@ public class Sonic3kObjectRegistry extends AbstractObjectRegistry {
                     }
                     return new AizGiantRideVineObjectInstance(spawn);
                 });
-        factories.put(Sonic3kObjectIds.MHZ_SWING_VINE,
+        factories.put(Sonic3kObjectIds.LBZ_TUBE_ELEVATOR,
                 (spawn, registry) -> {
                     S3kZoneSet zoneSet = getCurrentZoneSet();
                     if (zoneSet == S3kZoneSet.SKL && currentRomZoneId() == Sonic3kZoneIds.ZONE_MHZ) {
                         return new MhzSwingVineObjectInstance(spawn);
                     }
-                    return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                    if (zoneSet != S3kZoneSet.S3KL) {
+                        return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                    }
+                    return new LbzTubeElevatorInstance(spawn);
                 });
         factories.put(Sonic3kObjectIds.BREAKABLE_WALL,
                 (spawn, registry) -> new BreakableWallObjectInstance(spawn));
@@ -177,6 +181,14 @@ public class Sonic3kObjectRegistry extends AbstractObjectRegistry {
                     }
                     return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
                 });
+        factories.put(Sonic3kObjectIds.LBZ_RIDE_GRAPPLE,
+                (spawn, registry) -> {
+                    S3kZoneSet zoneSet = getCurrentZoneSet();
+                    if (zoneSet != S3kZoneSet.S3KL) {
+                        return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                    }
+                    return new LbzRideGrappleInstance(spawn);
+                });
         factories.put(Sonic3kObjectIds.AUTOMATIC_TUNNEL,
                 (spawn, registry) -> new AutomaticTunnelObjectInstance(spawn));
         factories.put(Sonic3kObjectIds.MHZ_MUSHROOM_CAP,
@@ -186,6 +198,14 @@ public class Sonic3kObjectRegistry extends AbstractObjectRegistry {
                         return new MhzMushroomCapObjectInstance(spawn);
                     }
                     return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                });
+        factories.put(Sonic3kObjectIds.LBZ_ROLLING_DRUM,
+                (spawn, registry) -> {
+                    S3kZoneSet zoneSet = getCurrentZoneSet();
+                    if (zoneSet != S3kZoneSet.S3KL) {
+                        return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                    }
+                    return new LbzRollingDrumInstance(spawn);
                 });
         factories.put(Sonic3kObjectIds.AUTO_SPIN,
                 (spawn, registry) -> new AutoSpinObjectInstance(spawn));
@@ -975,6 +995,14 @@ public class Sonic3kObjectRegistry extends AbstractObjectRegistry {
                         return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
                     }
                     return new OrbinautBadnikInstance(spawn);
+                });
+        factories.put(Sonic3kObjectIds.RIBOT,
+                (spawn, registry) -> {
+                    S3kZoneSet zoneSet = getCurrentZoneSet();
+                    if (zoneSet != S3kZoneSet.S3KL) {
+                        return new PlaceholderObjectInstance(spawn, getPrimaryName(spawn.objectId(), zoneSet));
+                    }
+                    return new RibotBadnikInstance(spawn);
                 });
         factories.put(Sonic3kObjectIds.FLYBOT_767,
                 (spawn, registry) -> {